requirements.txt
aiogram=3.0.0b5
python-dotenv=0.21.0